+++ src/Futureburn.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Futureburn.Core;
-
-public class Class1
-{
-
-}